fix: Add retry decorator to stochastic tests (#678)
